(parse) Port parser tests to Jest + snapshots
Since the parser produces an AST, it lends itself _very_ well to
snapshot testing.  Jest makes that quite convenient.
